sh-sci.c: Fix issue "exit" command doesn't work correctly
Function "sci_tx_dma_release" usually receives NULL value from variable "chan"
when "exit" command is used on UART console terminal. When passing "chan"
as a parameter or uses its field members, function "dmaengine_terminate_all",
"dma_unmap_single", and "dma_release_channel" will generate the below error:

|    Unable to handle kernel NULL pointer dereference at virtual address 00000000
|      pgd = c0003000
|      [00000000] *pgd=80000040004003, *pmd=00000000
|
|      Backtrace:
|      [<c02a9aa0>] (work_fn_tx) from [<c003999c>] (process_one_work+0x1d0/0x308)
|       r9:00000000 r8:c12eed00 r7:00000000 r6:c12e87c0 r5:c09c39d4 r4:c466df80
|      [<c00397cc>] (process_one_work) from [<c003a624>] (worker_thread+0x2c4/0x404)
|       r10:00000008 r9:c466df98 r8:c12e87d4 r7:c0958100 r6:c12e87c0 r5:c12e87c0
|       r4:c466df80
|      [<c003a360>] (worker_thread) from [<c003edb0>] (kthread+0xdc/0xf4)
|       r10:00000000 r9:00000000 r8:00000000 r7:c003a360 r6:c466df80 r5:00000000
|       r4:c45f3300
|      [<c003ecd4>] (kthread) from [<c0010088>] (ret_from_fork+0x14/0x2c)
|       r7:00000000 r6:00000000 r5:c003ecd4 r4:c45f3300

In conlusion, it's a good idea to call the above functions when "chan" is not NULL. Moreover,
the reason for removing the below message is that most of the time "exit" command is called,
terminal will display this message. So, it will impact user's experience.

|    dev_warn(port->dev, "Failed preparing Tx DMA descriptor\n");

Signed-off-by: Loc Pham <loc.pham.xh@rvc.renesas.com>
Signed-off-by: Nguyen Van Linh [FGA.BU13] <LinhNV25@fsoft.com.vn>
Signed-off-by: vietn <vietn@fsoft.com.vn>