sh-sci.c: Fix issue "exit" command doesn't work correctly
Function "sci_tx_dma_release" usually receives NULL value from variable
"chan" when "exit" command is used on UART console terminal. When passing
"chan" as a parameter or uses its field members, function
"dmaengine_terminate_all", "dma_unmap_single", and "dma_release_channel"
will generate the error.

In conlusion, it's a good idea to call the above functions when "chan" is
not NULL. Moreover, the reason for removing the below message is that most
of the time "exit" command is called, terminal will display this message.
So, it will impact user's experience.

|    dev_warn(port->dev, "Failed preparing Tx DMA descriptor\n");

Signed-off-by: Loc Pham <loc.pham.xh@rvc.renesas.com>
Signed-off-by: Nguyen Van Linh [FGA.BU13] <LinhNV25@fsoft.com.vn>
Signed-off-by: vietn <vietn@fsoft.com.vn>