drm: rcar-du: Fix mode_fixup callback function
commit a7847445b073da3ed533b645fd801cd49a17b67d
from git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-backport.git

It is prohibition to set up the width of the multiple of 16 for
the specification of H/W in R-Car series. Therefore, when it was
the width which is not a multiple of 16, the procedure which
returns an error was added.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Hung Dong <hung.dong.xd@rvc.renesas.com>