i2c: sh_mobile: fix clock calculation for r8a7744 SoC
r8a7744 SoC have so fast input clocks that the ICCL/H registers only
count every second clock to have a meaningful 9-bit range. The driver
was already prepared for that happening, but didn't use it so far.
Add the proper DT configuration for r8a7744 SoC that need it.

Signed-off-by: thongsyho <thong.ho.px@rvc.renesas.com>
Signed-off-by: Kha Tran <kha.tran.px@rvc.renesas.com>

Signed-off-by: Nguyen Van Linh [FGA.BU13] <LinhNv25@fsoft.com.vn>
Signed-off-by: vietn <vietn@fsoft.com.vn>

%% original patch: 0380-i2c-sh_mobile-fix-clock-calculation-for-r8a7744-SoC.patch