ARM: dts: r8a7742: change EthAVB status to disabled
This is of iWave u-boot workaround.
In u-boot, when SW0 is ON, VIN0 is selected but EthAVB status(in dtb) is not
changed. But when SW0 is OFF, AVB is selected and EthAVB status is changed to
"okay". So we must keep EthAVB to default status "disabled".
(VIN0 and EthAVB is muxed)

Signed-off-by: nhannguyen <nhan.nguyen.yb@renesas.com>
Signed-off-by: vietn <vietn@fsoft.com.vn>

%% original patch: 0579-ARM-dts-r8a7742-change-EthAVB-status-to-disabled.patch