linux-renesas:r8a7743-dtsi:add vcp0 vpc0 prr vpc1 vpc0xy vpc1xy for uvcs-kernel-module to support LTS4.4.6
Update by Huan Thai <huan.thai.xw@rvc.renesas.com>
In details:
These are issues when porting from 3.10 to 4.4
    Issue 1: Can not request interrupt for HW IP (VCP)
    Issue 2: Can not map HW IP register (VPC)
These is a difference in register platform device for the driver:
    - The old driver use kernel API (platform_device_register())
to register it device to kernel,
    - The new driver, device binding to driver is register by the new way
that use the device info from <target-board>.dtb file.

vpc1, vpc0xy, and vpc1xy are added as reserve only vcp0, vpc0 and prr is used.
The prr is used for getting LSI information which tell the driver
what kind of board it running

Signed-off-by: vietn <vietn@fsoft.com.vn>

%% original patch: 0033-linux-renesas-r8a7743-dtsi-add-vcp0-vpc0-prr-vpc1-vp.patch

Signed-off-by: Nguyen Van Linh [FGA.BU13] <LinhNV25@fsoft.com.vn>
Signed-off-by: vietn <vietn@fsoft.com.vn>

%% original patch: 0036-linux-renesas-r8a7743-dtsi-add-vcp0-vpc0-prr-vpc1-vp.patch