Revert "drm: rcar-du: Fix H/V sync signal polarity configuration"
This reverts commit 0d2b7767611fc1762140614452e21d558aa5e462.

We got issue du crash when doing IT, The failed cases are random and not
always. After comparing with cip3 which gave a good IT result, we found
that the commit 0d2b776761 made issue.

More investigation is needed, but we revert it for now.

Signed-off-by: Binh Nguyen <binh.nguyen.uw@renesas.com>
(cherry picked from commit 40bfd8cddaebf221e4061123695923102e0e9a70)
Signed-off-by: Binh Nguyen <binh.nguyen.uw@renesas.com>