clk: renesas: r8a77990: Correct parent clock of Audio-DMAC
The clock source of the AIX-bus clock (266.66 MHz) used for Audio-DMAC
DMA transfer is correct below:

R-Car H3	R-Car E3
Audio-DMAC	clock source
Audio-DMAC0	s1d2

As a result, changes the parent clocks of the Audio-DMAC0 module
on R-Car E3 to S1D2.

NOTE: This information will be reflected in the R-Car Gen3 Hardware
Manual in the future.

Signed-off-by: hienhuynh <hien.huynh.px@rvc.renenas.com>