gpio-rcar: Prevent to force the interrupt handler as threaded
On RT system, when kernel is booting with CONFIG_IRQ_FORCED_THREADING
flag, the gpio_rcar_irq_handler() will be forced as threaded and then
will generate a WARNING relating to interrupt handlers.

This happens because all of them invoke generic_handle_irq() from the
requsted handler. generic_handle_irq grabs raw_locks and this needs to
run in raw-irq context.

This commit add IRQF_NO_THREAD flag when requesting an interrupt to
prevent the warning.

Signed-off-by: Danh Duong <danhdc1@fsoft.com.vn>