thermal: rcar_gen3_thermal: Fix interrupts are not raised issue on E3
In current code, the thermal interrupts
are not raised when temperature is changed.

The cause is that, software reads temperature (CTEMP)
right after changing interrupt control register.
During that time, the hardware temperature is changing
and still unstable. It leads to wrong value on software.

This patch updates so that software checks
and waits for hardware changing become stably.

Signed-off-by: Huynh Thanh Hung <hung.huynh.wz@rvc.renesas.com>