mmc: tmio: Add selecting optimum TAP when all of TAP is OK
In SDHI manual rev.0.14 (Nov.11,2016), SCC_SMPCMP is added.
This patch improves TAP selection by using SCC_SMPCMP.
When all of TAP succeed in tuning, any TAP can be selected.
By referring to SCC_SMPCMP, it is possible to select the
optimum TAP when all of the TAP is OK.

Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>
[saito: rework commit title and message. original title is
"mmc: tmio: Add detecting a change point of data".]

Signed-off-by: Huynh Thanh Hung <hung.huynh.wz@rvc.renesas.com>