serial: sh-sci: Fix transfer sequence of unsupport DMA transfer
This patchs was created based on commit e9bad4f9ec44
("serial: sh-sci: Fix transfer sequence of unsupport DMA transfer").
In several R-Car SoC, the channel which is supporting the DMA transfer,
and the channel which is not supporting the DMA transfer are present
in (H)SCIx device.

On this SoC, when a DMA transfer of (H)SCIx device is enabled,
the following errors occur frequently:
sh-sci xxx: dma_request_slave_channel failed

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>
Signed-off-by: Lam Ngoc Son <son.lam.ym@rvc.renesas.com>