soc: renesas: rcar-sysc: Adjust the waiting time to cover the worst case
As description in HWM rev0.51E, 9.4 Usage notes, page 455:
"It takes several hundreds of microseconds to shutting off and
resuming power domain. Because actual time required
for shutting off and resuming depends on the status of on-board
power line, shutoff/resume time is not guaranteed by
electrical specification.
"
So, the safe value of waiting is about 1000us.

And number of non-CPU power domains is 10, so in the worst case,
the total waiting time is 10*1000us.

This patch adjust the waiting time to cover for the worst case.

Signed-off-by: Dien Pham <dien.pham.ry@rvc.renesas.com>
(cherry picked from commit 6e3e12d9c12a40d4ca465afe19c64bc6f9eb7de3)
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: hienhuynh <hien.huynh.px@rvc.renenas.com>