soc: renesas: rcar-sysc: Include 3DG as always-on power domains
As in commit (soc: renesas: rcar-sysc: Improve power
domains always on implementation (1*)),  it shows that
Hardware needs to be controlled with the state transitions
do not conflict for each power domains.

After verifying again, the patterns of conflict are updated as below.

1. CA5x  vs CA5x : Exclusion control with HW
2. CA5x  vs 3DG  : Exclusion control with SW (Not implemented yet)
3. CA5x  vs Other: Exclusion control with SW (Not implemented yet)
4. 3DG   vs Other: Exclusion control with SW (already implemented)
5. Other vs Other: Exclusion control with SW (already implemented)

In that,
- 3. has been implemented in (1*).
- 2. has not been implemented and this patch will implement for it.

Note: On Salvator-X, this pattern was only detected
when SW11-4 is OFF (debug ON mode.)

Signed-off-by: hienhuynh <hien.huynh.px@rvc.renenas.com>