media: rcar-vin: do not allow changing scaling and composing while streaming
It is possible on Gen2 to change the registers controlling composing and
scaling while the stream is running. It is however not a good idea to do
so and could result in trouble. There are also no good reasons to allow
this, remove immediate reflection in hardware registers from
vidioc_s_selection and only configure scaling and composing when the
stream starts.

Signed-off-by: Niklas Soderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Huy Tran <huy.tran.xd@rvc.renesas.com>