clk: renesas: r8a77{95,96,965}: Correct parent clock of SYS-DMAC
The clock source of the AIX BUS clock (266.66 MHz) used for SYS-DMAC
DMA transfer is correct below:

			R-Car H3	R-Car M3-W	R-Car M3-N
	sys-dmac ch	clock source	clock source	clock source
	============================================================
	SYS-DMAC0	s0d3		s0d3		s0d3
	SYS-DMAC1	s3d1		s3d1		s3d1
	SYS-DMAC2	s3d1		s3d1		s3d1

As a result, changes the parent clocks of the SYS-DMAC{1,2} module clocks
on R-Car H3, R-Car M3-W and R-Car M3-N to S3D1.

NOTE: This information will be reflected in the R-Car Gen3 Hardware
Manual in the future.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>