clk: renesas: r8a774c0: Correct RCLK handling
According to Hardware Manual, R8A774C0 SoC has the
RCLK Frequency Control Register (RCKCR), which determines the OSC
and RINT predivider values, and selection of the RCLK clock source
between RINT and the On-Chip Oscillator.

Hence change the OSC and RINT clock definitions to use the RCKCR
divider, and add the missing On-Chip Oscillator and RCLK clock
source switching logic.

Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>