soc: renesas: r8a774c0-sysc: Fix initialization order of 3DG-{A,B}
Reference upstream commit id:
<a64597227d0de5610407fcc95dc835745a53f5d0>

The workaround for the wrong hierarchy of the 3DG-{A,B} power
domains on RZ/G2E ES1.0 corrected the parent domains.
However, the 3DG-{A,B} power domains were still initialized and
powered in the wrong order, causing 3DG operation to fail.

Fix this by changing the order in the table at runtime, when running
on an affected SoC.

This work is based on the work done by Geert for R-Car E3.

Signed-off-by: Biju Das <biju.das@bp.renesas.com>
Signed-off-by: Hien Huynh <hien.huynh.px@renesas.com>