iommu/ipmmu-vmsa: Add override support for the actual number of MMU contexts
Add a IPMMU_VMSA_CTX_NUM to override the number of MMUs contexts that
you want to use.

The specified number of MMU contexts are used for the Renesas
IPMMU driver and the remaining contexts are unused.

SoCs, which is not R-Car Gen3, have no effect on IPMMU_VMSA_CTX_NUM.

Signed-off-by: Tomoharu Fukawa <tomoharu.fukawa.eb@renesas.com>
Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>
Signed-off-by: Hai Nguyen Pham <hai.pham.ud@renesas.com>
Signed-off-by: Chinh Nguyen <chinh.nguyen.jg@renesas.com>